requirements.txt
# Proxy
virtualenv
Flask
requests
setuptools
build
gunicorn
pyyaml
pycryptodome
backoff
# Azure
azure-mgmt-resource 
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-msi
azure-mgmt-authorization
azure-mgmt-subscription
azure-identity
azure-cosmos
azure-storage-blob
# GCP
google-cloud-compute
google-cloud-resource-manager
google-cloud-iam
google-cloud-logging
firebase-admin
